move stuff back to lib/state
