+++ erigon-lib/state/aggregator_bench_test.go
@@ -36,7 +36,6 @@ import (
 	"github.com/erigontech/erigon-lib/log/v3"
 	"github.com/erigontech/erigon-lib/recsplit"
 	"github.com/erigontech/erigon-lib/seg"
-	ee "github.com/erigontech/erigon-lib/state/entity_extras"
 )
 
 func testDbAndAggregatorBench(b *testing.B, aggStep uint64) (kv.RwDB, *Aggregator) {
@@ -98,7 +97,7 @@ func BenchmarkAggregator_Processing(b *testing.B) {
 }
 
 func queueKeys(ctx context.Context, seed, ofSize uint64) <-chan []byte {
-	rnd := NewRnd(seed)
+	rnd := newRnd(seed)
 	keys := make(chan []byte, 1)
 	go func() {
 		for {
@@ -117,22 +116,22 @@ func queueKeys(ctx context.Context, seed, ofSize uint64) <-chan []byte {
 
 func Benchmark_BtreeIndex_Search(b *testing.B) {
 	logger := log.New()
-	rnd := NewRnd(uint64(time.Now().UnixNano()))
+	rnd := newRnd(uint64(time.Now().UnixNano()))
 	tmp := b.TempDir()
 	defer os.RemoveAll(tmp)
 	dataPath := "../../data/storage.256-288.kv"
 
 	indexPath := filepath.Join(tmp, filepath.Base(dataPath)+".bti")
 	comp := seg.CompressKeys | seg.CompressVals
-	ee.BuildBtreeIndex(b, dataPath, indexPath, comp, 1, logger, true)
+	buildBtreeIndex(b, dataPath, indexPath, comp, 1, logger, true)
 
 	M := 1024
-	kv, bt, err := ee.OpenBtreeIndexAndDataFile(indexPath, dataPath, uint64(M), comp, false)
+	kv, bt, err := OpenBtreeIndexAndDataFile(indexPath, dataPath, uint64(M), comp, false)
 	require.NoError(b, err)
 	defer bt.Close()
 	defer kv.Close()
 
-	keys, err := ee.PivotKeysFromKV(dataPath)
+	keys, err := pivotKeysFromKV(dataPath)
 	require.NoError(b, err)
 	getter := seg.NewReader(kv.MakeGetter(), comp)
 
@@ -153,17 +152,17 @@ func benchInitBtreeIndex(b *testing.B, M uint64, compression seg.FileCompression
 	tmp := b.TempDir()
 	b.Cleanup(func() { os.RemoveAll(tmp) })
 
-	dataPath := ee.GenerateKV(b, tmp, 52, 10, 1000000, logger, 0)
+	dataPath := generateKV(b, tmp, 52, 10, 1000000, logger, 0)
 	indexPath := filepath.Join(tmp, filepath.Base(dataPath)+".bt")
 
-	ee.BuildBtreeIndex(b, dataPath, indexPath, compression, 1, logger, true)
+	buildBtreeIndex(b, dataPath, indexPath, compression, 1, logger, true)
 
-	kv, bt, err := ee.OpenBtreeIndexAndDataFile(indexPath, dataPath, M, compression, false)
+	kv, bt, err := OpenBtreeIndexAndDataFile(indexPath, dataPath, M, compression, false)
 	require.NoError(b, err)
 	b.Cleanup(func() { bt.Close() })
 	b.Cleanup(func() { kv.Close() })
 
-	keys, err := ee.PivotKeysFromKV(dataPath)
+	keys, err := pivotKeysFromKV(dataPath)
 	require.NoError(b, err)
 	return kv, bt, keys, dataPath
 }
@@ -172,7 +171,7 @@ func Benchmark_BTree_Seek(b *testing.B) {
 	M := uint64(1024)
 	compress := seg.CompressNone
 	kv, bt, keys, _ := benchInitBtreeIndex(b, M, compress)
-	rnd := NewRnd(uint64(time.Now().UnixNano()))
+	rnd := newRnd(uint64(time.Now().UnixNano()))
 	getter := seg.NewReader(kv.MakeGetter(), compress)
 
 	b.Run("seek_only", func(b *testing.B) {
@@ -229,7 +228,7 @@ func Benchmark_Recsplit_Find_ExternalFile(b *testing.B) {
 		b.Skip("requires existing KV index file at ../../data/storage.kv")
 	}
 
-	rnd := NewRnd(uint64(time.Now().UnixNano()))
+	rnd := newRnd(uint64(time.Now().UnixNano()))
 	tmp := b.TempDir()
 
 	defer os.RemoveAll(tmp)
@@ -245,7 +244,7 @@ func Benchmark_Recsplit_Find_ExternalFile(b *testing.B) {
 
 	getter := decomp.MakeGetter()
 
-	keys, err := ee.PivotKeysFromKV(dataPath)
+	keys, err := pivotKeysFromKV(dataPath)
 	require.NoError(b, err)
 
 	for i := 0; i < b.N; i++ {
@@ -398,10 +397,4 @@ func BenchmarkDb_BeginFiles_Throughput_IO(b *testing.B) {
 			tx.Rollback()
 		}
 	})
-}
-
-type RndGen = ee.RndGen
-
-func NewRnd(seed uint64) *ee.RndGen {
-	return ee.NewRnd(seed)
 }
