--- erigon-lib/state/entity_extras/files_item.go
@@ -1,13 +1,0 @@
-package entity_extras
-
-import (
-	"github.com/erigontech/erigon-lib/recsplit"
-	"github.com/erigontech/erigon-lib/seg"
-)
-
-type FilesItem interface {
-	Segment() *seg.Decompressor
-	AccessorIndex() *recsplit.Index
-	BtIndex() *BtIndex
-	ExistenceFilter() *ExistenceFilter
-}
