--- erigon-lib/state/entity_extras/testutils.go
@@ -1,134 +1,0 @@
-package entity_extras
-
-import (
-	"context"
-	"encoding/binary"
-	"fmt"
-	"math/rand/v2"
-	"path/filepath"
-	"testing"
-
-	randOld "math/rand"
-
-	"github.com/c2h5oh/datasize"
-	"github.com/erigontech/erigon-lib/common"
-	"github.com/erigontech/erigon-lib/common/background"
-	"github.com/erigontech/erigon-lib/etl"
-	"github.com/erigontech/erigon-lib/log/v3"
-	"github.com/erigontech/erigon-lib/seg"
-	"github.com/stretchr/testify/require"
-)
-
-type RndGen struct {
-	*rand.Rand
-	oldGen *randOld.Rand
-}
-
-func NewRnd(seed uint64) *RndGen {
-	return &RndGen{
-		Rand:   rand.New(rand.NewChaCha8([32]byte{byte(seed)})),
-		oldGen: randOld.New(randOld.NewSource(int64(seed))),
-	}
-}
-func (r *RndGen) IntN(n int) int                   { return int(r.Uint64N(uint64(n))) }
-func (r *RndGen) Read(p []byte) (n int, err error) { return r.oldGen.Read(p) } // seems `go1.22` doesn't have `Read` method on `math/v2` generator
-
-func GenerateKV(tb testing.TB, tmp string, keySize, valueSize, keyCount int, logger log.Logger, compressFlags seg.FileCompression) string {
-	tb.Helper()
-
-	rnd := NewRnd(0)
-	values := make([]byte, valueSize)
-
-	dataPath := filepath.Join(tmp, fmt.Sprintf("%dk.kv", keyCount/1000))
-	comp, err := seg.NewCompressor(context.Background(), "cmp", dataPath, tmp, seg.DefaultCfg, log.LvlDebug, logger)
-	require.NoError(tb, err)
-
-	bufSize := 8 * datasize.KB
-	if keyCount > 1000 { // windows CI can't handle much small parallel disk flush
-		bufSize = 1 * datasize.MB
-	}
-	collector := etl.NewCollector(BtreeLogPrefix+" genCompress", tb.TempDir(), etl.NewSortableBuffer(bufSize), logger)
-
-	for i := 0; i < keyCount; i++ {
-		key := make([]byte, keySize)
-		n, err := rnd.Read(key[:])
-		require.EqualValues(tb, keySize, n)
-		binary.BigEndian.PutUint64(key[keySize-8:], uint64(i))
-		require.NoError(tb, err)
-
-		n, err = rnd.Read(values[:rnd.IntN(valueSize)+1])
-		require.NoError(tb, err)
-
-		err = collector.Collect(key, values[:n])
-		require.NoError(tb, err)
-	}
-
-	writer := seg.NewWriter(comp, compressFlags)
-
-	loader := func(k, v []byte, _ etl.CurrentTableReader, _ etl.LoadNextFunc) error {
-		err = writer.AddWord(k)
-		require.NoError(tb, err)
-		err = writer.AddWord(v)
-		require.NoError(tb, err)
-		return nil
-	}
-
-	err = collector.Load(nil, "", loader, etl.TransformArgs{})
-	require.NoError(tb, err)
-
-	collector.Close()
-
-	err = comp.Compress()
-	require.NoError(tb, err)
-	comp.Close()
-
-	decomp, err := seg.NewDecompressor(dataPath)
-	require.NoError(tb, err)
-	defer decomp.Close()
-	compPath := decomp.FilePath()
-	ps := background.NewProgressSet()
-
-	IndexFile := filepath.Join(tmp, fmt.Sprintf("%dk.bt", keyCount/1000))
-	err = BuildBtreeIndexWithDecompressor(IndexFile, decomp, compressFlags, ps, tb.TempDir(), 777, logger, true)
-	require.NoError(tb, err)
-
-	return compPath
-}
-
-// Opens .kv at dataPath and generates index over it to file 'indexPath'
-func BuildBtreeIndex(tb testing.TB, dataPath, indexPath string, compressed seg.FileCompression, seed uint32, logger log.Logger, noFsync bool) {
-	tb.Helper()
-	decomp, err := seg.NewDecompressor(dataPath)
-	require.NoError(tb, err)
-	defer decomp.Close()
-
-	err = BuildBtreeIndexWithDecompressor(indexPath, decomp, compressed, background.NewProgressSet(), filepath.Dir(indexPath), seed, logger, noFsync)
-	require.NoError(tb, err)
-}
-
-// takes first 100k keys from file
-func PivotKeysFromKV(dataPath string) ([][]byte, error) {
-	decomp, err := seg.NewDecompressor(dataPath)
-	if err != nil {
-		return nil, err
-	}
-
-	getter := decomp.MakeGetter()
-	getter.Reset(0)
-
-	key := make([]byte, 0, 64)
-
-	listing := make([][]byte, 0, 1000)
-
-	for getter.HasNext() {
-		if len(listing) > 100000 {
-			break
-		}
-		key, _ := getter.Next(key[:0])
-		listing = append(listing, common.Copy(key))
-		getter.Skip()
-	}
-	decomp.Close()
-
-	return listing, nil
-}
